move test_setSelection into utils.ts
