+++ src/modules/editorStore/actions/notations/setColumnsNotation.test.ts
@@ -6,7 +6,7 @@ import { cleanup, renderHook } from '@testing-library/react';
 
 import { useEditorStore } from '../../useEditorStore';
 import { resetStore } from '../resets/resetStore';
-import { test_setSelection } from '../testUtils';
+import { test_setSelection } from '../utils';
 import { setColumnsNotation } from './setColumnsNotation';
 
 beforeEach(() => {
