+++ src/modules/editorStore/actions/utils.ts
@@ -1,4 +1,11 @@
 import { SelectionError, SelectionErrorCode, SelectionErrorCulprit } from '../errors/SelectionError';
+import { useEditorStore } from '../useEditorStore';
+
+export const test_setSelection = (section: number, start: number, end: number) => {
+	useEditorStore.setState((state) => {
+		state.currentSelection = { section, start, end };
+	});
+};
 
 export const validateSelection = (selection: ColumnSelection, tablature: Tablature): NonBlankColumnSelection => {
 	const { section, start, end } = selection;
